+++ CHANGELOG.md
@@ -1,1 +1,8 @@
-* Fix incompatibility with Create (and other mods) causing swords to deal no damage
+* Japanese translation by Mohuzato
+* Added new config options:
+* * **Client config**
+* * * disable particles from lit Bonfires
+* * * set the sound for lit Bonfires this can be any sound event or empty to disable
+* * **Common config**
+* * * disable setting the respawn point when using a Bonfire
+* * * disable Ash dropping from mobs